Create new namespace for import functions
Prior to this change, all import functions were contained inside the
Model class. This change introduces a new namespace FileMngr, that
groups the code used for file import. Currently, only the import
function for XYZ files is contained, but the remaining functions will
follow.
